Main Structure
Designed structure of server/client
Moved Terrain creation to subfolder
Cleaned up test files
